added a account app

+++ main/views.py
@@ -23,43 +23,74 @@ def ditail(request,id):
     
 #Add movies to database
 def add_movie(request):
-    if request.method == "POST":
-        form = MovieForm(request.POST or None) 
+    if request.user.is_authenticated:
+        if request.user.is_superuser:
         
-        
-       #check if form is valid or not 
-        if form.is_valid():
-            data = form.save(commit=False)
-            data.save()
-            return redirect('main:home')
-    else:
-        form = MovieForm()
-    return render(request, 'main/addmovies.html', {'form':form,
-                                                    'controller':"Add movies"})
     
+            if request.method == "POST":
+                form = MovieForm(request.POST or None) 
+                
+                
+               #check if form is valid or not 
+                if form.is_valid():
+                    data = form.save(commit=False)
+                    data.save()
+                    return redirect('main:home')
+            else:
+                form = MovieForm()
+            return render(request, 'main/addmovies.html', {'form':form,
+                                                            'controller':"Add movies"})
+        #if they are not admin
+        else:
+            return redirect('main:home')
+            
+    #if they are n login 
+    return redirect('accounts:login_user')
 
 #Edit movie
 def edit_movies(request,id):
-        #get the movie with the id
-        movie = Movie.objects.get(id=id)
+        if request.user.is_authenticated:
+            if request.user.is_superuser:
+                    #get the movie with the id
+                movie = Movie.objects.get(id=id)
+                    
+                    #form check
+                if request.method == "POST":
+                    form = MovieForm(request.POST or None,instance=movie)
+                        #check if form is valid
+                    if form.is_valid():
+                        data = form.save(commit=False)
+                        data.save()
+                        return redirect('main:ditail', id)
+                        
+                else:
+                    form = MovieForm(instance=movie)
+                return render(request, 'main/addmovies.html',{'form':form, 'controller':"Edit Movies"})
+                #if they are not admin
+            else:
+                return redirect('main:home')
+                
+        #if they are n login 
+        return redirect('accounts:login_user')
         
-        #form check
-        if request.method == "POST":
-            form = MovieForm(request.POST or None,instance=movie)
-            #check if form is valid
-            if form.is_valid():
-                data = form.save(commit=False)
-                data.save()
-                return redirect('main:ditail', id)
-            
-        else:
-            form = MovieForm(instance=movie)
-        return render(request, 'main/addmovies.html',{'form':form, 'controller':"Edit Movies"})
         
  #delete movie post    
 def delete_movie(request,id):
-    movie = Movie.objects.get(id=id)
+    if request.user.is_authenticated:
+        if request.user.is_superuser:
+
+
+            movie = Movie.objects.get(id=id)
+                
+                #delete the movie -object
+            movie.delete()
+            return redirect('main:home')
     
-    #delete the movie -object
-    movie.delete()
-    return redirect('main:home')
+    
+    
+            #if they are not admin
+        else:
+            return redirect('main:home')
+            
+    #if they are n login 
+    return redirect('accounts:login_user')
